Fix FlxText and problem with camera zoom
The problem was not related to FlxText only, but any FlxSprite. If the zoom was much greater than 1x (e.g 1.5x or 2x), objects positioned near the border of the screen would not be displayed.

It was happening because FlxCamera did not adjust its bounds when the zoom changes. As a consequence, if the camera had a zoom of 2x, the bounds would allow it to see just the elements within the 1x rectangle. The problem does not happen when the zoom is <= 1x.

I fixed the problem by adding a "zoom offset" to the bounds when checking them.