Tweaks to make code more robust
FlxG.cameraRemove() now uses FlxG._game.contains() to check if the camera exists in the display list (better than catching exceptions).

FlxCamera.destroy() checks if the "screen" property was not already destroyed before it tries to nullify it. It prevents a crash when FlxG.removeCamera() tries to remove and destroy an already removed (and destroyed) camera.